requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
requests==2.28.2
beautifulsoup4==4.11.2
django-phonenumber-field==7.0.2
phonenumbers==8.13.6
django-rest-framework==0.1.0
